handle goods detail

+++ platform-system/platform-system-api/src/main/java/com/klauting/timormall/system/api/service/IGoodsBasicService.java
@@ -17,7 +17,7 @@ public interface IGoodsBasicService {
 	
 	GoodsBasic queryObject(Long id);
 	
-	GoodsBasicDto queryGoodsIdAndPlatformId(Long goodsId,Long platformId);
+	GoodsBasicDto queryGoodsIdAndPlatformId(Long goodsId,int platformId);
 	
 	
 	List<GoodsBasicDto> queryGoodsBasicDtoList(Map<String, Object> map);
